requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
requests
beautifulsoup4
bcrypt
pyjwt
pandas
python-multipart
celery
redis